requirements.txt
flask>=3.0.0
lxml>=5.0.0
anthropic>=0.18.0
openai>=1.0.0
boto3>=1.34.0
openpyxl>=3.1.0
requests>=2.31.0
Pillow>=10.0.0
